+++ docs/release-notes.rst
@@ -24,8 +24,9 @@ The client and server are released together.  The client is deployed as a jar fi
 Maven Nexus repository. The server is deployed as a Docker image to a Docker registry.
 
 Version 2.0.0, 14 Nov 2018
------------------=--------
+--------------------------
 
+* Add method to get user unread notification count (`ACUMOS-1883 <https://jira.acumos.org/browse/ACUMOS-1883>`_)
 * Revise restricted-federation query method to use page parameter
 * Requires database schema version 2.0